reagganging get_DB functions in net instead of sync, linting
