+++ fusion_report/common/net.py
@@ -7,6 +7,9 @@ import ssl
 import urllib.error
 import urllib.request
 import time
+import pandas as pd
+from zipfile import ZipFile
+
 
 from argparse import Namespace
 from typing import List
@@ -17,7 +20,10 @@ from fusion_report.common.exceptions.download import DownloadException
 from fusion_report.common.logger import Logger
 from fusion_report.data.cosmic import CosmicDB
 from fusion_report.settings import Settings
-
+from multiprocessing import Pool
+from fusion_report.data.fusiongdb import FusionGDB
+from fusion_report.data.fusiongdb2 import FusionGDB2
+from fusion_report.data.mitelman import MitelmanDB
 
 class Net:
 
@@ -89,6 +95,53 @@ class Net:
             db.setup(files, delimiter='\t', skip_header=True)
         except urllib.error.HTTPError as ex:
             return_err.append(f'{Settings.COSMIC["NAME"]}: {ex}')
+
+    @staticmethod
+    def get_fusiongdb(self, return_err: List[str]) -> None:
+        """Method for download FusionGDB database."""
+
+        pool_params = [
+            (f'{Settings.FUSIONGDB["HOSTNAME"]}/{x}', True) for x in Settings.FUSIONGDB["FILES"]
+        ]
+        pool = Pool(Settings.THREAD_NUM)
+        pool.starmap(Net.get_large_file, pool_params)
+        pool.close()
+        pool.join()
+        db = FusionGDB('.')
+        db.setup(Settings.FUSIONGDB['FILES'], delimiter='\t', skip_header=False)
+
+    @staticmethod
+    def get_fusiongdb2(self, return_err: List[str]) -> None:
+        """Method for download FusionGDB2 database."""
+        try:
+            url: str = f'{Settings.FUSIONGDB2["HOSTNAME"]}/{Settings.FUSIONGDB2["FILE"]}'
+            Net.get_large_file(url)
+            file: str = f'{Settings.FUSIONGDB2["FILE"]}'
+            df = pd.read_excel(file)
+            df["fusion"] = df["5'-gene (text format)"] + "--" + df["3'-gene (text format)"]
+            file_csv = 'fusionGDB2.csv'
+            df['fusion'].to_csv(file_csv, header=False, index=False, sep=',', encoding='utf-8')
+
+            db = FusionGDB2('.')
+            db.setup([file_csv], delimiter=',', skip_header=False)
+
+        except DownloadException as ex:
+            return_err.append(f'FusionGDB2: {ex}')
+
+    @staticmethod
+    def get_mitelman(self, return_err: List[str]) -> None:
+        """Method for download Mitelman database."""
+        try:
+            url: str = f'{Settings.MITELMAN["HOSTNAME"]}/{Settings.MITELMAN["FILE"]}'
+            Net.get_large_file(url)
+            with ZipFile(Settings.MITELMAN['FILE'], 'r') as archive:
+                files = [x for x in archive.namelist() if "mitelman_db/MBCA.TXT.DATA" in x]
+                archive.extractall()
+
+            db = MitelmanDB('.')
+            db.setup(files, delimiter='\t', skip_header=False, encoding='ISO-8859-1')
+        except DownloadException as ex:
+            return_err.append(f'Mitelman: {ex}')
 
     @staticmethod
     def clean():
