+++ fusion_report/download.py
@@ -7,7 +7,6 @@ from typing import List
 from fusion_report.common.exceptions.download import DownloadException
 from fusion_report.common.logger import Logger
 from fusion_report.common.net import Net
-from fusion_report.sync import Sync
 
 class Download:
     """Class designed for downloading any type of required database.
@@ -36,13 +35,13 @@ class Download:
         os.chdir(params.output)
 
         # MITELMAN
-        Sync.get_mitelman(self, return_err)
+        Net.get_mitelman(self, return_err)
 
         # FusionGDB
-        Sync.get_fusiongdb(self, return_err)
+        Net.get_fusiongdb(self, return_err)
 
         # FusionGDB2
-        Sync.get_fusiongdb2(self, return_err)
+        Net.get_fusiongdb2(self, return_err)
 
         # COSMIC
         Net.get_cosmic(self.cosmic_token, return_err)
